+++ static/css/main.css
@@ -1,3 +1,4 @@
+
 .img-preview {
     width: 256px;
     height: 256px;
