+++ static/css/styles.css
@@ -1,0 +1,131 @@
+/* Existing styles */
+:root {
+    --white: #F5F5F5;
+    --light: #E0C9CB;
+    --tan: #C3A1A0;
+    --pink: #D9AAB7;
+    --rose: #BA7E7E;
+    --dark: #4E4E4E;
+  }
+
+html {
+    display: grid;
+    min-height: 100%;
+}
+
+body {
+    display: grid;
+    background: var(--light);
+    font-family: 'Lato' , 'sans-serif';
+    text-transform: uppercase;
+}
+
+.container {
+    position: relative;
+    margin: auto;
+    overflow: hidden;
+    width: 1300px;
+    height: 700px;
+    background:var(--white);
+    box-shadow: 5px 5px 15px rgba(var(--rose), .5);
+    border-radius: 10px;
+    display: flex;
+    align-items: center;
+}
+
+/*.productCard {
+    position: relative;
+  margin: auto;
+  overflow: hidden;
+  width: 520px;
+  height: 350px;
+  background: #f5f5f5;
+  box-shadow: 5px 5px 15px rgba(#d603af, .5);
+  border-radius: 10px;
+}*/
+
+img {
+    width: 600px;
+    margin-left: 60px;
+}
+
+.productInfo {
+    position: absolute;
+    width: 40%;
+    height: 100%;
+    top: 18%;
+    left: 60%;
+}
+
+.query {
+    font-size: 2em;
+    color: var(--dark);
+    letter-spacing: 1px;
+}
+
+.productTitle {
+    font-size: 3em;
+    color: var(--dark);
+    margin-top: -5px;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+}
+
+.productPrice {
+    color: var(--dark);
+    margin-top: -5pxpx;
+}
+
+.productLink {
+    background: var(--light);
+    padding: 10px;
+    display: inline-block;
+    outline: 0;
+    border: 0;
+    text-align: center;
+    width: 60%;
+    margin: -1px;
+    border-radius: 2px;
+    text-transform: uppercase;
+    letter-spacing: 1px;
+    color: var(--white);
+    cursor: pointer;
+    text-decoration: none;
+}
+
+.productLink:hover {
+    background: var(--rose);
+    transition: all 0.1s ease-in-out;
+}
+
+.navigation {
+    position: absolute;
+    top: 50%;
+    transform: translateY(-50%);
+    display: flex;
+    justify-content: space-between; 
+    width: 99%;
+}
+
+/* Existing styles for arrows */
+
+.nav-btn {
+    background: #f1f1f1;
+    border: none;
+    border-radius: 5px;
+    color: #333;
+    font-size: 21px;
+    cursor: pointer;
+    z-index: 1;
+}
+
+/*#prevBtn {
+    left: 10px;
+}
+
+#nextBtn {
+    right: 10px;
+}*/
+
+
