+++ static/js/main.js
@@ -47,6 +47,7 @@ $(document).ready(function () {
                 $('#result').fadeIn(600);
                 $('#result').text(' Result:  ' + data);
                 console.log('Success!');
+                window.location.href = '/shower';
             },
         });
     });
