+++ templates/index.html
@@ -1,8 +1,8 @@
 {% extends "base.html" %} {% block content %}
 
-<h2>Image Identifier</h2>
+<h2>ShopWiseAI</h2>
 
-<div>
+<div class="unique">
     <form id="upload-file" method="post" enctype="multipart/form-data">
         <label for="imageUpload" class="upload-label">
             Choose...
@@ -16,7 +16,7 @@
             </div>
         </div>
         <div>
-            <button type="button" class="btn btn-primary btn-lg " id="btn-predict">Predict!</button>
+            <button type="button" class="btn btn-primary btn-lg " id="btn-predict">Upload</button>
         </div>
     </div>
