+++ templates/index2.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Your Webpage</title>
+    <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400&display=swap" rel="stylesheet">
+    <script src="https://code.jquery.com/jquery-3.6.4.min.js"></script>
+    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Lato:400,700">
+    <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
+</head>
+<body>
+<div class="part_2">
+<div id="scraper_results" class="container">
+    <!-- Cards will be dynamically added here -->
+</div>
+
+<div class="navigation">
+    <button id="prevBtn" class="nav-btn" onclick="plusSlides(-1)">&#10094;</button>
+    <button id="nextBtn" class="nav-btn" onclick="plusSlides(1)">&#10095;</button>
+</div>
+</div>
+
+<script src="{{ url_for('static', filename='js/script.js') }}"></script>
+
+</body>
+</html>
