+++ templates/login.html
@@ -29,5 +29,5 @@
         </form>
     </div>
 </body>
-<script src="{{ url_for('static', filename='js/script1.js') }}"></script>
+<script src="{{ url_for('static', filename='js/login.js') }}"></script>
 </html>
