+++ templates/register.html
@@ -39,5 +39,5 @@
         </form>
     </div>
 </body>
-<script src="{{ url_for('static', filename='js/script.js') }}"></script>
+<script src="{{ url_for('static', filename='js/register.js') }}"></script>
 </html>
